reorganize things to simplify a bit
