--- featuretestapp-all/featuretestapp/src/main/java/org/webpieces/app/base/libs/RemoteService.java
@@ -1,10 +1,0 @@
-package org.webpieces.app.base.libs;
-
-import java.util.concurrent.CompletableFuture;
-
-public interface RemoteService {
-
-	public CompletableFuture<Integer> fetchRemoteValue(String s, int i);
-	
-	public void sendData(int num);
-}
