--- featuretestapp-all/featuretestapp/src/main/java/org/webpieces/app/base/libs/SomeLibrary.java
@@ -1,7 +1,0 @@
-package org.webpieces.app.base.libs;
-
-public interface SomeLibrary {
-
-	public void doSomething(int something);
-
-}
