--- featuretestapp-all/featuretestapp/src/main/java/org/webpieces/app/base/libs/SomeLibraryImpl.java
@@ -1,9 +1,0 @@
-package org.webpieces.app.base.libs;
-
-public class SomeLibraryImpl implements SomeLibrary {
-
-	@Override
-	public void doSomething(int something)  {
-	}
-
-}
