+++ featuretestapp-all/featuretestapp/src/main/java/org/webpieces/app/example1/business/Hydrator.java
@@ -2,8 +2,8 @@ package org.webpieces.app.example1.business;
 
 import javax.inject.Inject;
 
-import org.webpieces.app.example1.model.Tweet;
-import org.webpieces.app.example1.remoteclients.HydratorService;
+import org.webpieces.app.example1.business.model.Tweet;
+import org.webpieces.app.example1.business.remoteclients.HydratorService;
 
 //public class HydratorImpl implements Hydrator {
 public class Hydrator {
