+++ featuretestapp-all/featuretestapp/src/main/java/org/webpieces/app/example1/business/TweetIdSearch.java
@@ -4,8 +4,8 @@ import java.util.List;
 
 import javax.inject.Inject;
 
-import org.webpieces.app.example1.remoteclients.TweetSearchService;
-import org.webpieces.app.example1.routes.SearchRequest;
+import org.webpieces.app.example1.business.remoteclients.TweetSearchService;
+import org.webpieces.app.example1.SearchRequest;
 
 public class TweetIdSearch {
   private TweetSearchService tweetSearchService;
