+++ featuretestapp-all/featuretestapp/src/main/java/org/webpieces/app/example1/business/UserSearch.java
@@ -2,8 +2,8 @@ package org.webpieces.app.example1.business;
 
 import javax.inject.Inject;
 
-import org.webpieces.app.example1.model.User;
-import org.webpieces.app.example1.remoteclients.UserSearchService;
+import org.webpieces.app.example1.business.model.User;
+import org.webpieces.app.example1.business.remoteclients.UserSearchService;
 
 //public class UserSearchImpl implements UserSearch {
 public class UserSearch {
