+++ featuretestapp-all/featuretestapp/src/main/java/org/webpieces/app/example1/business/remoteclients/HydratorService.java
@@ -1,0 +1,7 @@
+package org.webpieces.app.example1.business.remoteclients;
+
+import org.webpieces.app.example1.business.model.Tweet;
+
+public interface HydratorService {
+  Tweet hydrate(Integer tweetId);
+}
