+++ featuretestapp-all/featuretestapp/src/main/java/org/webpieces/app/example1/business/remoteclients/UserSearchService.java
@@ -1,0 +1,7 @@
+package org.webpieces.app.example1.business.remoteclients;
+
+import org.webpieces.app.example1.business.model.User;
+
+public interface UserSearchService {
+  User lookup(String userId);
+}
