+++ featuretestapp-all/featuretestapp/src/main/java/org/webpieces/app/example1/business/remoteclients/impl/TweetSearchServiceImpl.java
@@ -1,0 +1,16 @@
+package org.webpieces.app.example1.business.remoteclients.impl;
+
+import org.webpieces.app.example1.business.remoteclients.TweetSearchService;
+
+import java.util.ArrayList;
+import java.util.List;
+
+public class TweetSearchServiceImpl implements TweetSearchService {
+  @Override
+  public List<Integer> query(String queryString) {
+    List<Integer> list = new ArrayList<>();
+    list.add(56);
+    list.add(72);
+    return list;
+  }
+}
