+++ featuretestapp-all/featuretestapp/src/main/java/org/webpieces/app/example1/business/remoteclients/impl/UserSearchServiceImpl.java
@@ -1,0 +1,11 @@
+package org.webpieces.app.example1.business.remoteclients.impl;
+
+import org.webpieces.app.example1.business.model.User;
+import org.webpieces.app.example1.business.remoteclients.UserSearchService;
+
+public class UserSearchServiceImpl implements UserSearchService {
+  @Override
+  public User lookup(String userId) {
+    return new User("Patrick", "Patrick Stover");
+  }
+}
