--- featuretestapp-all/featuretestapp/src/main/java/org/webpieces/app/example1/exception/AuthenticationException.java
@@ -1,7 +1,0 @@
-package org.webpieces.app.example1.exception;
-
-public class AuthenticationException extends RuntimeException {
-  public AuthenticationException(String message) {
-    super(message);
-  }
-}
