--- featuretestapp-all/featuretestapp/src/main/java/org/webpieces/app/example1/exception/InternalServerException.java
@@ -1,11 +1,0 @@
-package org.webpieces.app.example1.exception;
-
-public class InternalServerException extends RuntimeException {
-  public InternalServerException(String message) {
-    super(message);
-  }
-
-  public InternalServerException(Exception cause) {
-    super(cause);
-  }
-}
