--- featuretestapp-all/featuretestapp/src/main/java/org/webpieces/app/example1/jsonIndex.html
@@ -1,1 +1,0 @@
-This is the basic /json start page.
