--- featuretestapp-all/featuretestapp/src/main/java/org/webpieces/app/example1/remoteclients/HydratorService.java
@@ -1,7 +1,0 @@
-package org.webpieces.app.example1.remoteclients;
-
-import org.webpieces.app.example1.model.Tweet;
-
-public interface HydratorService {
-  Tweet hydrate(Integer tweetId);
-}
