--- featuretestapp-all/featuretestapp/src/main/java/org/webpieces/app/example1/remoteclients/UserSearchService.java
@@ -1,7 +1,0 @@
-package org.webpieces.app.example1.remoteclients;
-
-import org.webpieces.app.example1.model.User;
-
-public interface UserSearchService {
-  User lookup(String userId);
-}
