--- featuretestapp-all/featuretestapp/src/main/java/org/webpieces/app/example1/search.html
@@ -1,13 +1,0 @@
-#{if tweets}#
-  Here are your tweets: ${tweets}$
-#{/if}#
-#{else}#
-  No tweets found!
-#{/else}#
-
-#{if users}#
-  Here are your users: ${users}$
-#{/if}#
-#{else}#
-  No users found!
-#{/else}#
