+++ featuretestapp-all/featuretestapp/src/test/java/org/webpieces/app/example1/TestSearchRequestResponse.java
@@ -11,15 +11,12 @@ import org.junit.Test;
 import org.webpieces.JavaCache;
 import org.webpieces.app.Server;
 import org.webpieces.app.ServerConfig;
-import org.webpieces.app.example1.remoteclients.HydratorService;
-import org.webpieces.app.example1.remoteclients.TweetSearchService;
-import org.webpieces.app.example1.remoteclients.UserSearchService;
-import org.webpieces.app.example1.routes.SearchRequest;
+import org.webpieces.app.example1.business.remoteclients.HydratorService;
+import org.webpieces.app.example1.business.remoteclients.TweetSearchService;
 import org.webpieces.app.example1.util.Responses;
 import org.webpieces.app.mock.MockHydratorService;
 import org.webpieces.app.mock.MockRemoteSystem;
 import org.webpieces.app.mock.MockTweetSearchService;
-import org.webpieces.app.mock.MockUserSearchService;
 import org.webpieces.data.api.DataWrapper;
 import org.webpieces.data.api.DataWrapperGenerator;
 import org.webpieces.data.api.DataWrapperGeneratorFactory;
