+++ featuretestapp-all/featuretestapp/src/test/java/org/webpieces/app/mock/MockTweetSearchService.java
@@ -1,10 +1,8 @@
 package org.webpieces.app.mock;
 
-import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.List;
 
-import org.webpieces.app.example1.remoteclients.TweetSearchService;
+import org.webpieces.app.example1.business.remoteclients.TweetSearchService;
 
 public class MockTweetSearchService implements TweetSearchService {
   private List<Integer> tweetIds;
